% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.R
\name{gbm}
\alias{gbm}
\title{Simulation for Geometric Brownian Motion}
\usage{
gbm(mu, sigma, x0, t, n_sims = 100, n_steps = 100, method = "exact")
}
\arguments{
\item{mu}{numeric. The drift parameter.}

\item{sigma}{numeric. The diffusion parameter.}

\item{x0}{numeric. The initial value.}

\item{t}{numeric. The time to simulate the process. Default \code{1}.}

\item{n_sims}{integer. The number of simulations. Default \code{100}.}

\item{n_steps}{integer. The number of steps to simulate the process. Default \code{100}.}

\item{method}{character. The method used to simulate the process. Default \code{exact}. Options include \code{"exact"}, \code{"euler"}, and \code{"milstein"}.}
}
\value{
An S4 Class object \code{"Simulations"} containing the information of new simulations.
}
\description{
Simulation for Geometric Brownian Motion
}
\examples{
gbm(mu = 0.1, sigma = 0.2, x0 = 100, t = 1)
gbm(mu = 0.1, sigma = 0.2, x0 = 100, t = 1, n_sims = 1000, n_steps = 1000)
gbm(mu = 0.1, sigma = 0.2, x0 = 100, t = 1, method = "euler")
}
