% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\docType{class}
\name{Simulations-class}
\alias{Simulations-class}
\title{S4 Class of Stochastic Process Simulations}
\value{
An S4 Class object containing the information of new simulations.
}
\description{
S4 Class of Stochastic Process Simulations
}
\section{Slots}{

\describe{
\item{\code{process}}{character. The type of process the object represented.}

\item{\code{paths}}{numeric. A matrix containing the paths of the simulated stochastic process. Each row represent a step in time, and each column represents an individual path.}

\item{\code{method}}{character.}
}}

\examples{
# PLACEHOLDER
}
